Separated function definition from function call
